Added basic support for Analogizer
Added support for the Analogizer: analog video modes: RGBS, RGsB, YPbPr and SNAC.
Issue: In SMB3 in menu and map screens YPbPr works Ok, but when Mario/Luigui enters to level screen screen goes black. Returning to map or title screen and image recovers.
this could be due to something related to a change in video mode during game phases.