Added Chip32 loader and fixed some things to make room
At the end I'm forced to change some things in the Analogizer side to leave some logic resources shifting to BRAM when is possible. Also I've shorten the vector width in case sentences to the minimal. the last thing was to remove synchronizer for the interact.json settings, only keep sync the external_reset signal to avoid glitches on this vital signal.