General resources usage reduction to fit all
I have further reduced the resources consumed by the Analogizer framework by taking advantage of the fact that the video output is 18bits instead of 24bits. I have also added a TCL script and the Windows Powershell launcher file to generate the different bitstreams from the command line. The PLL used for NTSC/PAL timing is the same (nes_pll) but the clock speed settings are adjusted based on a parameter (USE_PAL_PLL). See the build_and_install_instructions.txt file for more information on this and the generate.tcl script.